rugs-mcp/requirements.txt
# MCP Server
mcp>=1.0.0

# HTTP client
httpx>=0.27.0

# Async support
anyio>=4.0.0

# Environment
python-dotenv>=1.0.0
